Move the SQS Queue controller's tests to use the new app context
